WIP for Hermitian metrics
I don't think this is going anywhere. It seems like we need the Chern
connection to zero out at the origin to be able to get the estimates we
need. If I want to keep doing this I should probably start with the
Kahler case and see if I can recreate Hitchin's results.